vetted_test_backend/requirements.txt
certifi==2018.8.24
chardet==3.0.4
diff-match-patch==20121119
Django==2.1.1
django-cors-headers==2.1.0
django-environ==0.4.4
django-jet==1.0.7
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
jet==1.0
odfpy==1.3.6
openpyxl==2.5.6
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
